v1.0.0: general code hygiene
