+++ app/src/main/res/values-night/colors.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="special_item">#1B0836</color>
+    <color name="regular_item">#ff303030</color>
 </resources>
